Create publish_trusted_testers.yml
Might have to wait in-between updating and publishing the item, not sure. Also need to retrieve some values from secrets. The token should probably be the refresh token and use it to get a new access token, though not sure if this requires that we check if there's an old one still valid...